+++ lwjson/src/include/lwjson/lwjson.h
@@ -29,7 +29,7 @@
  * This file is part of LwJSON - Lightweight JSON format parser.
  *
  * Author:          Tilen MAJERLE <tilen@majerle.eu>
- * Version:         v1.6.1
+ * Version:         v1.7.0
  */
 #ifndef LWJSON_HDR_H
 #define LWJSON_HDR_H
@@ -178,7 +178,7 @@ typedef enum {
     LWJSON_STREAM_STATE_PARSING_STRING, /*!< Parse string primitive */
     LWJSON_STREAM_STATE_PARSING_PRIMITIVE, /*!< Parse any primitive that is non-string, either "true", "false", "null" or a number */
     LWJSON_STREAM_STATE_EXPECTING_COMMA_OR_END, /*!< Expecting ',', '}' or ']' */
-    LWJSON_STREAM_STATE_EXPECTING_COLON, /*!< Expecting ':' */
+    LWJSON_STREAM_STATE_EXPECTING_COLON,        /*!< Expecting ':' */
 } lwjson_stream_state_t;
 
 /* Forward declaration */