+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## Develop
 
+## 1.7.0
+
 - Add clang-tidy
 - Add helper functions for sequence check in stream parsing
 - Add support to discard invalid JSON stream
+- Fixed some invalid JSON parsing in the streaming module
 
 ## 1.6.1